put dynamicComponents list in its own file
